+++ docs/reference/search/point-in-time-api.asciidoc
@@ -1,5 +1,3 @@
-[role="xpack"]
-[testenv="basic"]
 [[point-in-time-api]]
 === Point in time API
 ++++
