+++ rest-api-spec/src/main/resources/rest-api-spec/test/search/350_point_in_time.yml
@@ -37,8 +37,8 @@ setup:
 ---
 "basic":
   - skip:
-      version: " - 7.9.99"
-      reason: "point in time is introduced in 7.10"
+      version: " - 7.12.99"
+      reason: "point in time is introduced in server in 7.13"
   - do:
       open_point_in_time:
         index: test
@@ -47,7 +47,6 @@ setup:
 
   - do:
       search:
-        rest_total_hits_as_int: true
         body:
           size: 1
           query:
@@ -58,7 +57,7 @@ setup:
             id: "$point_in_time_id"
             keep_alive: 1m
 
-  - match: {hits.total: 3 }
+  - match: {hits.total.value: 3 }
   - length: {hits.hits: 1 }
   - match: {hits.hits.0._index: test }
   - match: {hits.hits.0._id: "172" }
@@ -77,7 +76,6 @@ setup:
   # search with a point in time
   - do:
       search:
-        rest_total_hits_as_int: true
         body:
           size: 1
           query:
@@ -88,7 +86,7 @@ setup:
           pit:
             id: "$point_in_time_id"
 
-  - match: {hits.total: 3 }
+  - match: {hits.total.value: 3 }
   - length: {hits.hits: 1 }
   - match: {hits.hits.0._index: test }
   - match: {hits.hits.0._id: "42" }
@@ -97,7 +95,6 @@ setup:
 
   - do:
       search:
-        rest_total_hits_as_int: true
         body:
           size: 1
           query:
@@ -108,7 +105,7 @@ setup:
           pit:
             id: "$point_in_time_id"
 
-  - match: {hits.total: 3}
+  - match: {hits.total.value: 3}
   - length: {hits.hits: 1 }
   - match: {hits.hits.0._index: test }
   - match: {hits.hits.0._id: "1" }
@@ -117,7 +114,6 @@ setup:
 
   - do:
       search:
-        rest_total_hits_as_int: true
         body:
           size: 1
           query:
@@ -129,7 +125,7 @@ setup:
             id: "$point_in_time_id"
             keep_alive: 1m
 
-  - match: {hits.total: 3}
+  - match: {hits.total.value: 3}
   - length: {hits.hits: 0 }
 
   - do:
@@ -150,7 +146,6 @@ setup:
 
   - do:
       search:
-        rest_total_hits_as_int: true
         body:
           size: 2
           query:
@@ -161,12 +156,50 @@ setup:
             id: "$point_in_time_id"
             keep_alive: 1m
 
-  - match: {hits.total: 4 }
+  - match: {hits.total.value: 4 }
   - length: {hits.hits: 2 }
   - match: {hits.hits.0._index: test }
   - match: {hits.hits.0._id: "172" }
   - match: {hits.hits.1._index: test2 }
   - match: {hits.hits.1._id: "45" }
+
+  - do:
+      close_point_in_time:
+        body:
+          id: "$point_in_time_id"
+
+---
+"msearch":
+  - skip:
+      version: " - 7.99.99"
+      reason: "After backport: 7.9.99 => point in time is introduced in 7.10"
+  - do:
+      open_point_in_time:
+        index: "t*"
+        keep_alive: 5m
+  - set: {id: point_in_time_id}
+
+  - do:
+      msearch:
+        body:
+          - {}
+          - { query: { match: { _index: test }}, pit: { id: "$point_in_time_id" }}
+          - {}
+          - { query: { match_all: {}}, pit: { id: "$point_in_time_id" }}
+
+  - match:  { responses.0.hits.total.value:     3   }
+  - match:  { responses.0.hits.total.relation:  eq  }
+  - match:  { responses.1.hits.total.value:     4   }
+  - match:  { responses.1.hits.total.relation:  eq  }
+
+  - do:
+      catch: bad_request
+      msearch:
+        body:
+          - index: index
+          - { query: { match: { foo: bar }}, pit: { id: "$point_in_time_id" }}
+          - index: index2
+          - { query: { match_all: {}}, pit: { id: "$point_in_time_id" }}
 
   - do:
       close_point_in_time:
