+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/XPackClientPlugin.java
@@ -208,10 +208,8 @@ import org.elasticsearch.xpack.core.rollup.action.StopRollupJobAction;
 import org.elasticsearch.xpack.core.rollup.job.RollupJob;
 import org.elasticsearch.xpack.core.rollup.job.RollupJobStatus;
 import org.elasticsearch.xpack.core.runtimefields.RuntimeFieldsFeatureSetUsage;
-import org.elasticsearch.xpack.core.search.action.ClosePointInTimeAction;
 import org.elasticsearch.xpack.core.search.action.GetAsyncSearchAction;
 import org.elasticsearch.xpack.core.search.action.GetAsyncStatusAction;
-import org.elasticsearch.xpack.core.search.action.OpenPointInTimeAction;
 import org.elasticsearch.xpack.core.search.action.SubmitAsyncSearchAction;
 import org.elasticsearch.xpack.core.searchablesnapshots.SearchableSnapshotFeatureSetUsage;
 import org.elasticsearch.xpack.core.security.SecurityFeatureSetUsage;
@@ -505,9 +503,6 @@ public class XPackClientPlugin extends Plugin implements ActionPlugin, NetworkPl
                 GetAsyncSearchAction.INSTANCE,
                 GetAsyncStatusAction.INSTANCE,
                 DeleteAsyncResultAction.INSTANCE,
-                // Point in time
-                OpenPointInTimeAction.INSTANCE,
-                ClosePointInTimeAction.INSTANCE,
                 // Data streams,
                 CreateDataStreamAction.INSTANCE,
                 GetDataStreamAction.INSTANCE,
